new constructor for lens
will probably still need to change.  also trying to evaluate best approach to group visibility.

need to be able to flexibly modify groups but also need to maintain the promise that groups never overlap, may need a few more member functions, or to just say fuckit and let it be public with a little pinky promise that users wont break that guarentee